fix: revert autoFormatExpiry to use "input" event instead of "keyup"
For some reason, despite everything being used having support for the OS and browsers we are testing against, using the "keyup" event means certain key strokes weren't working properly in the selenium tests - typically in one of, but never in all of, Safari/Edge/Windows XP. Key strokes such as ARROW_LEFT or HOME to get to the start of the expiry date input were resulting in the cursor remaining at the end of the input (on occasion, one of the ARROW_LEFT instructions would work as expected, but very inconsistently).

Switching back to "input" event solved this issue.